BigDecimal for price & total, removed redundant code async service

+++ src/main/java/com/manage/order/dto/OrderResponseDTO.java
@@ -1,5 +1,6 @@
 package com.manage.order.dto;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -10,5 +11,5 @@ public record OrderResponseDTO(
     Integer customerId,
     String customerName,
     String status,
-    Double total
+    BigDecimal total
 ) {}
